Removed AboutMe, added ColorCode
I ended up removing AboutMe as I realized for the aspect of this website that's my resume and about me, I can just add that to the main home page when you first enter the website.

The ColorCode section is something i really want to implement where essentially you upload a bunch of items and they show up on one side and you can add them to a chart. Then it'll color organize the chart (or maybe I can assort it using other information).